feat
pequena troca para melhorar outro modulo

+++ gestao.h
@@ -4,7 +4,7 @@ struct gestao{
     char nomep[50];
     float valor;
     char descricaop[1000];
-    float quantidade;
+    int quantidade;
     char status;
 };
